% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_forest_study.R
\name{add_to_forest}
\alias{add_to_forest}
\title{Add a list of R objects to an existing forest_study object.}
\usage{
add_to_forest(
  x,
  forest_name,
  data_type = "auto",
  engine = arrow::write_parquet,
  engine_args = NULL,
  file_ext = NULL
)
}
\arguments{
\item{x}{\verb{Named list} of objects to create an arrow-compatible data set from.}

\item{forest_name}{\verb{Character scalar}. Name of the database folder that
should be created.}

\item{data_type, file_ext}{\verb{Character scalar}. Specify file suffix and
file extension, respectively. (Default: \code{NULL})}

\item{engine}{\code{function} The function that will write out the content of \code{x}.
(Default: \code{arrow::write_parquet})}

\item{engine_args}{\code{list} of other arguments to \code{engine} function argument.
(Default: NULL)}
}
\description{
Add a list of R objects to an existing forest_study object.
}
