% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_unpack_repack.R
\name{assemble_SE}
\alias{assemble_SE}
\title{Assemble a SummarizedExperiment from structured component files.}
\usage{
assemble_SE(x)
}
\arguments{
\item{x}{\verb{Character scalar} path to the file.}
}
\description{
Assemble a SummarizedExperiment from structured component files.
}
