% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a data set from a list of objects.}
\usage{
create_dataset(
  x,
  dataset_name,
  engine = arrow::write_parquet,
  ...,
  .file_suffix = NULL,
  .file_ext = NULL
)
}
\arguments{
\item{x}{\verb{Named list} of objects to create an arrow-compatible dataset from}

\item{dataset_name}{\verb{Character scalar}. Name of the database folder that
should be created.}

\item{engine}{\code{function} The function that will write out the content of \code{x}.
(Default: \code{arrow::write_parquet})}

\item{...}{Additional arguments to function specified with 'engine' argument.}

\item{.file_suffix, file_ext}{\verb{Character scalar}. Specify file suffix and
file extension, respectively. (Default: \code{NULL})}
}
\description{
Meant to be seamlessly compatible with \code{arrow::read_parquet()}.
}
