% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_unpack_repack.R
\name{disassemble_SE}
\alias{disassemble_SE}
\title{Disassemble a SummarizedExperiment into component tables and save them}
\usage{
disassemble_SE(x, path = "flatSE")
}
\arguments{
\item{x}{\code{SummarizedExperiment}}

\item{path}{\verb{Character scalar} path to write to.}
}
\description{
Disassemble a SummarizedExperiment into component tables and save them
}
