% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{load_data_type}
\alias{load_data_type}
\title{#' @title Create a dataset
#' @export
#' @param x an R object.
#' @examples
#' create_study_forest
create_study_forest <- S7::new_generic("create_study_forest", "x")}
\usage{
load_data_type(x, data_type, ...)
}
\description{
#' @title Create a dataset
#' @export
#' @param x an R object.
#' @examples
#' create_study_forest
create_study_forest <- S7::new_generic("create_study_forest", "x")
}
