% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{macro_study}
\alias{macro_study}
\title{An S7 object to organise a meta-analysis}
\usage{
macro_study(x)
}
\arguments{
\item{x}{\verb{Character scalar} pointing to the relevant folder.}
}
\description{
An S7 object to organise a meta-analysis
}
