% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_methods.R
\name{random_forest_study}
\alias{random_forest_study}
\title{Randomly generate a forest_study and corresponding folders.}
\usage{
random_forest_study(
  n_trees = 10,
  n_data_slots = 10,
  sparseness = 0.5,
  forest_name = tempfile()
)
}
\arguments{
\item{n_trees}{\verb{Numeric scalar} number of studies to generate}

\item{forest_name}{\verb{Character scalar}. Name of the database folder that
should be created.}
}
\description{
Mostly for instructive and debugging purposes.
}
\examples{
x <- random_forest_study(n_trees = 100)
x

}
