% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_methods.R
\name{random_macro_study}
\alias{random_macro_study}
\title{Randomly generate a macro_study and corresponding folders.}
\usage{
random_macro_study(n, dataset_name = tempfile())
}
\arguments{
\item{n}{\verb{Numeric scalar} number of studies to generate}

\item{dataset_name}{\verb{Character scalar}. Name of the database folder that
should be created.}
}
\description{
Mostly for instructive and debugging purposes.
}
\examples{
x <- random_macro_study(n = 100)
x

}
