% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{meta_study}
\alias{meta_study}
\title{An S7 object to organised a meta-analysis}
\usage{
meta_study(x)
}
\arguments{
\item{x}{\verb{Character scalar} pointing to the relevant folder.}
}
\description{
An S7 object to organised a meta-analysis
}
\examples{

meta_study

}
