% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{meta_study}
\alias{meta_study}
\alias{new_meta_study}
\title{An S7 object to organise a meta-analysis}
\usage{
meta_study(path = character(0))

new_meta_study(x)
}
\arguments{
\item{x}{\verb{Character scalar} pointing to the relevant folder.}
}
\description{
An S7 object to organise a meta-analysis
}
