% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_methods.R
\name{random_meta_study}
\alias{random_meta_study}
\title{Randomly generate a meta_study and corresponding folders.}
\usage{
random_meta_study(n, dataset_name = tempdir())
}
\arguments{
\item{n}{\verb{Numeric scalar} number of studies to generate}

\item{dataset_name}{\verb{Character scalar}. Name of the database folder that
should be created.}
}
\description{
Mostly for instructive and debugging purposes.
}
\examples{
x <- random_meta_study(n = 100)
x

}
